+++ gemfiles/Gemfile.rails-6.0.rb
@@ -2,19 +2,14 @@ source 'https://rubygems.org'
 
 gemspec path: '../'
 
-gem 'rails', '~> 4.2.0.rc3' do
-  gem 'activerecord'
-  gem 'railties'
-end
-
-gem 'i18n', '0.7.0.beta1'
+gem 'activerecord', '~> 6.0.0'
+gem 'railties', '~> 6.0.0'
 
 # Database Configuration
 group :development, :test do
   platforms :jruby do
-    gem 'activerecord-jdbcsqlite3-adapter', '>= 1.3.0.beta2'
-    gem 'activerecord-jdbcmysql-adapter', '>= 1.3.0.beta2'
-    gem 'activerecord-jdbcpostgresql-adapter', '>= 1.3.0.beta2'
+    gem 'activerecord-jdbcmysql-adapter', '~> 51.1'
+    gem 'activerecord-jdbcpostgresql-adapter', '~> 51.1'
     gem 'kramdown'
   end