+++ lib/friendly_id/history.rb
@@ -41,12 +41,12 @@ method.
       ...
 
       def find_post
-        @post = Post.find params[:id]
+        @post = Post.friendly.find params[:id]
 
         # If an old id or a numeric id was used to find the record, then
-        # the request path will not match the post_path, and we should do
-        # a 301 redirect that uses the current friendly id.
-        if request.path != post_path(@post)
+        # the request slug will not match the current slug, and we should do
+        # a 301 redirect to the new path
+        if params[:id] != @post.slug
           return redirect_to @post, :status => :moved_permanently
         end
       end
@@ -54,25 +54,27 @@ method.
 =end
   module History
 
+    module Configuration
+      def dependent_value
+        dependent.nil? ? :destroy : dependent
+      end
+    end
+
     def self.setup(model_class)
       model_class.instance_eval do
         friendly_id_config.use :slugged
+        friendly_id_config.class.send :include, History::Configuration
         friendly_id_config.finder_methods = FriendlyId::History::FinderMethods
-        if friendly_id_config.uses? :finders
-          relation.class.send(:include, friendly_id_config.finder_methods)
-          if ActiveRecord::VERSION::MAJOR == 4 && ActiveRecord::VERSION::MINOR == 2
-            model_class.send(:extend, friendly_id_config.finder_methods)
-          end
-        end
+        FriendlyId::Finders.setup(model_class) if friendly_id_config.uses? :finders
       end
     end
 
     # Configures the model instance to use the History add-on.
     def self.included(model_class)
       model_class.class_eval do
-        has_many :slugs, -> {order("#{Slug.quoted_table_name}.id DESC")}, {
+        has_many :slugs, -> {order(id: :desc)}, **{
           :as         => :sluggable,
-          :dependent  => :destroy,
+          :dependent  => @friendly_id_config.dependent_value,
           :class_name => Slug.to_s
         }
 
@@ -84,18 +86,17 @@ method.
       include ::FriendlyId::FinderMethods
 
       def exists_by_friendly_id?(id)
-        joins(:slugs).where(arel_table[friendly_id_config.query_field].eq(id)).exists? || joins(:slugs).where(slug_history_clause(id)).exists?
+        super || joins(:slugs).where(slug_history_clause(id)).exists?
       end
 
       private
 
       def first_by_friendly_id(id)
-        matching_record = where(friendly_id_config.query_field => id).first
-        matching_record || slug_table_record(id)
+        super || slug_table_record(id)
       end
 
       def slug_table_record(id)
-        select(quoted_table_name + '.*').joins(:slugs).where(slug_history_clause(id)).first
+        select(quoted_table_name + '.*').joins(:slugs).where(slug_history_clause(id)).order(Slug.arel_table[:id].desc).first
       end
 
       def slug_history_clause(id)
@@ -109,9 +110,10 @@ method.
     # to be conflicts. This will allow a record to revert to a previously
     # used slug.
     def scope_for_slug_generator
-      relation = super
-      return relation if new_record?
-      relation = relation.merge(Slug.where('sluggable_id <> ?', id))
+      relation = super.joins(:slugs)
+      unless new_record?
+        relation = relation.merge(Slug.where('sluggable_id <> ?', id))
+      end
       if friendly_id_config.uses?(:scoped)
         relation = relation.where(Slug.arel_table[:scope].eq(serialized_scope))
       end
@@ -120,17 +122,26 @@ method.
 
     def create_slug
       return unless friendly_id
-      return if slugs.first.try(:slug) == friendly_id
+      return if history_is_up_to_date?
       # Allow reversion back to a previously used slug
       relation = slugs.where(:slug => friendly_id)
       if friendly_id_config.uses?(:scoped)
         relation = relation.where(:scope => serialized_scope)
       end
-      relation.delete_all
+      relation.destroy_all unless relation.empty?
       slugs.create! do |record|
         record.slug = friendly_id
         record.scope = serialized_scope if friendly_id_config.uses?(:scoped)
       end
+    end
+
+    def history_is_up_to_date?
+      latest_history = slugs.first
+      check = latest_history.try(:slug) == friendly_id
+      if friendly_id_config.uses?(:scoped)
+        check = check && latest_history.scope == serialized_scope
+      end
+      check
     end
   end
 end