+++ lib/friendly_id/object_utils.rb
@@ -1,4 +1,16 @@
 module FriendlyId
+  # Instances of these classes will never be considered a friendly id.
+  # @see FriendlyId::ObjectUtils#friendly_id
+  UNFRIENDLY_CLASSES = [
+    Array,
+    FalseClass,
+    Hash,
+    NilClass,
+    Numeric,
+    Symbol,
+    TrueClass
+  ]
+
   # Utility methods for determining whether any object is a friendly id.
   #
   # Monkey-patching Object is a somewhat extreme measure not to be taken lightly
@@ -27,13 +39,7 @@ module FriendlyId
     #     "123".friendly_id?                #=> nil
     #     "abc123".friendly_id?             #=> true
     def friendly_id?
-      # Considered unfriendly if this is an instance of an unfriendly class or
-      # one of its descendants.
-      if FriendlyId::UNFRIENDLY_CLASSES.detect {|klass| self.class <= klass}
-        false
-      elsif respond_to?(:to_i) && to_i.to_s != to_s
-        true
-      end
+      true if respond_to?(:to_i) && to_i.to_s != to_s
     end
 
     # True if the id is definitely unfriendly, false if definitely friendly,
@@ -42,6 +48,29 @@ module FriendlyId
       val = friendly_id? ; !val unless val.nil?
     end
   end
+
+  module UnfriendlyUtils
+    def friendly_id?
+      false
+    end
+
+    def unfriendly_id?
+      true
+    end
+  end
+
+  def self.mark_as_unfriendly(klass)
+    klass.send(:include, FriendlyId::UnfriendlyUtils)
+  end
 end
 
 Object.send :include, FriendlyId::ObjectUtils
+
+# Considered unfriendly if object is an instance of an unfriendly class or
+# one of its descendants.
+
+FriendlyId::UNFRIENDLY_CLASSES.each { |klass| FriendlyId.mark_as_unfriendly(klass) }
+
+ActiveSupport.on_load(:active_record) do
+  FriendlyId.mark_as_unfriendly(ActiveRecord::Base)
+end