+++ test/candidates_test.rb
@@ -1,6 +1,6 @@
 require "helper"
 
-class CandidatesTest < Minitest::Test
+class CandidatesTest < TestCaseClass
 
   include FriendlyId::Test
 
@@ -111,6 +111,32 @@ class CandidatesTest < Minitest::Test
     end
     with_instances_of klass do |_, city|
       assert_equal "new-york-ewr", city.friendly_id
+    end
+  end
+
+  test "allows to iterate through candidates without passing block" do
+    klass = Class.new model_class do
+      def slug_candidates
+        :name
+      end
+    end
+    with_instances_of klass do |_, city|
+      candidates = FriendlyId::Candidates.new(city, city.slug_candidates)
+      assert_equal candidates.each, ['new-york']
+    end
+  end
+
+  test "iterates through candidates with passed block" do
+    klass = Class.new model_class do
+      def slug_candidates
+        :name
+      end
+    end
+    with_instances_of klass do |_, city|
+      collected_candidates = []
+      candidates = FriendlyId::Candidates.new(city, city.slug_candidates)
+      candidates.each { |candidate| collected_candidates << candidate }
+      assert_equal collected_candidates, ['new-york']
     end
   end