+++ test/reserved_test.rb
@@ -1,6 +1,6 @@
 require "helper"
 
-class ReservedTest < Minitest::Test
+class ReservedTest < TestCaseClass
 
   include FriendlyId::Test
 
@@ -59,6 +59,16 @@ class ReservedTest < Minitest::Test
         ["edit", "new"]
       end
       assert_raises(ActiveRecord::RecordInvalid) {record.save!}
+    end
+  end
+
+  test "should optionally treat reserved words as conflict" do
+    klass = Class.new(model_class) do
+      friendly_id :slug_candidates, :use => [:slugged, :reserved], :reserved_words => %w(new edit), :treat_reserved_as_conflict => true
+    end
+
+    with_instance_of(klass, name: 'new') do |record|
+      assert_match(/new-([0-9a-z]+\-){4}[0-9a-z]+\z/, record.slug)
     end
   end