+++ test/scoped_test.rb
@@ -20,7 +20,7 @@ class Publisher < ActiveRecord::Base
   has_many :novels
 end
 
-class ScopedTest < Minitest::Test
+class ScopedTest < TestCaseClass
 
   include FriendlyId::Test
   include FriendlyId::Test::Shared::Core
@@ -78,6 +78,19 @@ class ScopedTest < Minitest::Test
       record.slug = nil
       record.save!
       assert_equal old_id, record.friendly_id
+    end
+  end
+
+  test "should generate new slug when scope changes" do
+    transaction do
+      novelist = Novelist.create! :name => "a"
+      publisher = Publisher.create! :name => "b"
+      novel1 = Novel.create! :name => "c", :novelist => novelist, :publisher => publisher
+      novel2 = Novel.create! :name => "c", :novelist => novelist, :publisher => Publisher.create(:name => "d")
+      assert_equal novel1.friendly_id, novel2.friendly_id
+      novel2.publisher = publisher
+      novel2.save!
+      assert novel2.friendly_id != novel1.friendly_id
     end
   end